Move in-memory impl and test to inmemory folder
Move in-memory impl and test to inmemory folder

#3101
